spark/requirements.txt
# Delta Lake
delta-spark==3.3.0

# Amazon Deeque
pydeequ==1.4.0